+++ src/App/store.spec.js
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
-import Store from '../store';
+import Store from './store';
 
 const mockApps = [
   { id: '123', name: '123', type: 'local' },
